Separate operations from Cursor
