+++ lib/cursor.memory.js
@@ -5,7 +5,7 @@ const util = require('util');
 const common = require('metarhia-common');
 
 const constants = require('./constants');
-const transformations = require('./transformations');
+const operations = require('./operations');
 const Cursor = require('./cursor');
 
 function MemoryCursor(provider, dataset) {
@@ -198,7 +198,7 @@ MemoryCursor.prototype.fetch = function(done) {
   done = common.once(done);
   let dataset = common.clone(this.dataset);
   this.jsql.forEach(operation => {
-    const fn = MemoryCursor.operations[operation.op];
+    const fn = operations[operation.op];
     dataset = fn(operation, dataset);
   });
   done(null, dataset);
@@ -211,94 +211,5 @@ MemoryCursor.prototype.next = function(done) {
   done(new Error(constants.NOT_IMPLEMENTED));
   return this;
 };
-
-MemoryCursor.operations = {};
-
-MemoryCursor.operations.select = (operation, dataset) => {
-  const query = operation.query;
-  const fields = Object.keys(query);
-  const compare = transformations.compare;
-  return dataset.filter(record => {
-    let i, field, condition, keep;
-    for (i = 0; i < fields.length; i++) {
-      field = fields[i];
-      condition = query[field];
-      keep = compare(record, field, condition);
-      if (!keep) return false;
-    }
-    return true;
-  });
-};
-
-MemoryCursor.operations.distinct = (operation, dataset) => {
-  const keys = new Set();
-  let fields = operation.fields;
-  if (typeof(operation) === 'string') fields = [fields];
-  return dataset.filter((record) => {
-    const cols = fields || Object.keys(record).sort();
-    const key = cols.map(field => record[field]).join('\x00');
-    const has = keys.has(key);
-    keys.add(key);
-    return !has;
-  });
-};
-
-MemoryCursor.operations.order = (operation, dataset) => {
-  let fields = operation.fields;
-  if (typeof(operation) === 'string') fields = [fields];
-  dataset.sort((r1, r2) => {
-    const a1 = fields.map(field => r1[field]).join('\x00');
-    const a2 = fields.map(field => r2[field]).join('\x00');
-    if (a1 < a2) return -1;
-    if (a1 > a2) return 1;
-    return 0;
-  });
-  return dataset;
-};
-
-MemoryCursor.operations.desc = (operation, dataset) => {
-  let fields = operation.fields;
-  if (typeof(operation) === 'string') fields = [fields];
-  dataset.sort((r1, r2) => {
-    const a1 = fields.map(field => r1[field]).join('\x00');
-    const a2 = fields.map(field => r2[field]).join('\x00');
-    if (a1 < a2) return 1;
-    if (a1 > a2) return -1;
-    return 0;
-  });
-  return dataset;
-};
-
-MemoryCursor.operations.projection = (operation, dataset) => (
-  transformations.projection(operation.fields, dataset)
-);
-
-MemoryCursor.operations.row = (operation, dataset) => (
-  transformations.row(dataset)
-);
-
-MemoryCursor.operations.col = (operation, dataset) => (
-  transformations.col(dataset, operation.field)
-);
-
-MemoryCursor.operations.one = (operation, dataset) => (
-  dataset[0]
-);
-
-MemoryCursor.operations.union = (operation, dataset) => (
-  transformations.union(dataset, operation.cursor.dataset)
-);
-
-MemoryCursor.operations.intersection = (operation, dataset) => (
-  transformations.intersection(dataset, operation.cursor.dataset)
-);
-
-MemoryCursor.operations.difference = (operation, dataset) => (
-  transformations.difference(dataset, operation.cursor.dataset)
-);
-
-MemoryCursor.operations.complement = (operation, dataset) => (
-  transformations.complement(dataset, operation.cursor.dataset)
-);
 
 module.exports = MemoryCursor;
